<feat>[#283-WHS] 1001

+++ back-end-app/src/main/java/project/app/c109/backendapp/product/domain/dto/request/ProductRequestDTO.java
@@ -6,7 +6,7 @@ public class ProductRequestDTO {
     private Integer productSeq;
     private String productName;
     private Integer productPrice;
-    private Integer productDcrate;
+    private Double productDcrate;
     private String productInfo;
     private String productExp;
     private String productImage;
@@ -42,11 +42,11 @@ public class ProductRequestDTO {
         this.productPrice = productPrice;
     }
 
-    public Integer getProductDcrate() {
+    public Double getProductDcrate() {
         return productDcrate;
     }
 
-    public void setProductDcrate(Integer productDcrate) {
+    public void setProductDcrate(Double productDcrate) {
         this.productDcrate = productDcrate;
     }
